backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   groq
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
beautifulsoup4==4.14.3
    # via backend (pyproject.toml)
cachetools==6.2.3
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
distro==1.9.0
    # via
    #   groq
    #   openai
fastapi==0.124.4
    # via backend (pyproject.toml)
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via backend (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.3.0
    # via playwright
groq==0.37.1
    # via backend (pyproject.toml)
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   backend (pyproject.toml)
    #   groq
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.12.0
    # via openai
openai==2.11.0
    # via backend (pyproject.toml)
playwright==1.57.0
    # via backend (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   google-generativeai
    #   groq
    #   openai
pydantic-core==2.41.5
    # via pydantic
pyee==13.0.0
    # via playwright
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.2.1
    # via backend (pyproject.toml)
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via
    #   groq
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sse-starlette==3.0.3
    # via backend (pyproject.toml)
starlette==0.50.0
    # via fastapi
tqdm==4.67.1
    # via
    #   google-generativeai
    #   openai
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   fastapi
    #   google-generativeai
    #   groq
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via requests
uvicorn==0.38.0
    # via backend (pyproject.toml)
